requirements.txt
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
google-api-core==1.24.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
numpy==1.19.4
pandas==1.1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rsa==4.6
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.2
